+++ src/lib/functions/try/initializers/combine.ts
@@ -4,13 +4,34 @@ import {runInTry} from "../helpers";
 
 
 
-export async function combine<T extends any[], R>(...args: [...{ [K in keyof T]: Try<T[K]> }, (...values: T) => R]): Promise<Result>{
+export async function combine<T extends any[], R>(...args: [...{ [K in keyof T]: Try<T[K]> }, (...values: T) => R]): Promise<Result>;
+export async function combine<T extends any[], R>(...args: [...{ [K in keyof T]: Try<T[K]> }, (...values: T) => R, boolean]): Promise<Result>;
+export async function combine<T extends any[], R>(...args: any[]): Promise<Result>{
     const result = new Result();
-    const tries = args.slice(0, -1) as { [K in keyof T]: Try<T[K]> };
-    const func = args[args.length - 1] as (...values: T) => R;
-    const values: unknown[] = [];
+    // Check if the last argument is a boolean (parallel flag)
+    const hasParallelFlag = typeof args[args.length - 1] === 'boolean';
+    const tries = args.slice(0, hasParallelFlag ? -2 : -1) as { [K in keyof T]: Try<T[K]> };
+    const func = args[hasParallelFlag ? args.length - 2 : args.length - 1] as (...values: T) => R;
+    const parallel = hasParallelFlag ? (args[args.length - 1] as boolean) : true;
+    let values: unknown[] = [];
 
-    for(const v of tries){
+    const success = await runInTry(async () => {
+        if(parallel){
+            values =  await Promise.all(tries.map( values => values.get()))
+        }else{
+            for(const v of tries){
+                values.push(await v.get());
+            }
+        }
+    }, result)
+
+    if(!success)
+        return result;
+
+    //@ts-ignore
+    return result.setValue(await func(...values));
+
+    /* for(const v of tries){
         const success = await runInTry(async ()=>{
             values.push(await v.get());
         }, result);
@@ -21,6 +42,6 @@ export async function combine<T extends any[], R>(...args: [...{ [K in keyof T]:
     }
 
     //@ts-ignore
-    return result.setValue(await func(...values));
+    return result.setValue(await func(...values)); */
 
 }