+++ src/lib/functions/try/initializers/sequence.ts
@@ -4,19 +4,22 @@ import {runInTry} from "../helpers";
 
 
 
-export async function sequence<T extends readonly unknown[]>(tries: { [K in keyof T]: Try<T[K]> }): Promise<Result>{
+export async function sequence<T extends readonly unknown[]>(tries: { [K in keyof T]: Try<T[K]> }, parallel = false): Promise<Result>{
     const result = new Result();
-    const values: unknown[] = [];
+    let values: unknown[] = [];
 
-    for(const v of tries){
-        const success = await runInTry(async ()=>{
-            values.push(await v.get());
-        }, result);
+    const success = await runInTry(async () => {
+        if(parallel){
+            values =  await Promise.all(tries.map( values => values.get()))
+        }else{
+            for(const v of tries){
+                values.push(await v.get());
+            }
+        }
+    }, result)
 
-        if(!success)
-            return result;
-
-    }
+    if(!success)
+        return result;
 
     //@ts-ignore
     return result.setValue(values);