+++ src/test/Try.test.ts
@@ -505,7 +505,7 @@ describe("Try", () => {
 
     describe("Try.combine", () => {
 
-      test("combine should run all Try instances and pass the results to the provided function", async () => {
+      test("combine should resolve all Try instances sequentially and pass the results to the provided function", async () => {
           const r = Try.success(2);
           const r2 = Try.success(3);
           const r3 = Try.of(() => {
@@ -516,14 +516,14 @@ describe("Try", () => {
 
           const f = (a: number, b: number, c: string) => a + b + c;
 
-          const r4 = Try.combine(r, r2, r3, f);
+          const r4 = Try.combine(r, r2, r3, f, false);
 
           await expect(r4.get()).resolves.toBe("53");
           expect(r4.isSuccess()).toBe(true);
 
       });
 
-      test("combine should run all Try instances and results in Failure for the first instance that is a Failure", async () => {
+      test("combine should resolve all Try instances sequentially and results in Failure for the first instance that is a Failure", async () => {
           const r = Try.success(2);
           const r2 = Try.success(3);
           const r3 = Try.of(() => {
@@ -534,12 +534,68 @@ describe("Try", () => {
 
           const f = (a: number, b: number, c: string) => a + b + c;
 
-          const r4 = Try.combine(r, r2, r3, f);
+          const r4 = Try.combine(r, r2, r3, f, false);
 
           await expect(r4.get()).rejects.toThrow("Random error");
           expect(r4.isFailure()).toBe(true);
 
       });
+
+      test("combine should resolve all Try instances in parallel and pass the results to the provided function", async () => {
+        const r = Try.success(3).map(async v => {
+            await new Promise(resolve => setTimeout(resolve, 1000 * v));
+            return v;
+        });
+        const r2 = Try.success(4).map(async v => {
+            await new Promise(resolve => setTimeout(resolve, 1000  * v));
+            return v;
+        });
+        const r3 = Try.of(() => {
+            if(0.6 > 0.5) return "3";
+            throw new Error("Random error");
+        });
+
+
+        const f = (a: number, b: number, c: string) => a + b + c;
+
+        const r4 = Try.combine(r, r2, r3, f, true);
+
+        const start = performance.now();
+        await expect(r4.get()).resolves.toBe("73");
+        expect(r4.isSuccess()).toBe(true);
+        const timeTaken = (performance.now() - start) / 1000;
+        expect(timeTaken).toBeLessThan(5); //The maximum amount of time it should take is 5 seconds, because the longest running task is 4 seconds. The other 1 second is for the overhead of the test.
+
+      });
+
+      test("combine should resolve all Try instance in parallel and results in Failure for the first instance that is a Failure", async () => {
+        const r = Try.success(4).map(async v => {
+            await new Promise(resolve => setTimeout(resolve, 1000 * v));
+            return v;
+        });
+        const r2 = Try.success(3).map(async v => {
+            await new Promise(resolve => setTimeout(resolve, 1000  * v));
+            return v;
+        });
+        const r3 = Try.of(async () => {
+            if(0.3 > 0.5) await new Promise(resolve => setTimeout(resolve, 1000  * 2));
+            throw new Error("Random error 2");
+        });
+
+
+        const f = (a: number, b: number, c: string) => a + b + c;
+
+        const r4 = Try.combine(r, r2, r3, f);
+
+        const start = performance.now();
+        await expect(r4.get()).rejects.toThrow("Random error 2");
+        expect(r4.isFailure()).toBe(true);
+        const timeTaken = (performance.now() - start) / 1000;
+        expect(timeTaken).toBeLessThan(3); //The maximum amount of time it should take is 3 seconds, because the longest running task until the error is thrown is 2 seconds. The other 1 second is for the overhead of the test.
+
+    });
+
+
     });
 
     describe("Try.sequence", () => {
@@ -550,7 +606,7 @@ describe("Try", () => {
             const r3 = Try.success("4");
 
 
-            const r4 = Try.sequence([r, r2, r3]);
+            const r4 = Try.sequence([r, r2, r3], false);
 
             await expect(r4.get()).resolves.toEqual([2, 3, "4"]);
             expect(r4.isSuccess()).toBe(true);
@@ -566,11 +622,63 @@ describe("Try", () => {
             });
 
 
-            const r4 = Try.sequence([r, r2, r3]);
+            const r4 = Try.sequence([r, r2, r3], false);
 
             await expect(r4.get()).rejects.toThrow("Random error");
             expect(r4.isFailure()).toBe(true);
 
         });
+
+
+        test("sequence should run all Try instances in parallel and return an array of the result", async () => {
+            const r = Try.success(4).map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000 * v));
+                return v;
+            });
+            const r2 = Try.success(3).map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000  * v));
+                return v;
+            });
+            const r3 = Try.success("2")
+            .map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000  * Number(v)));
+                return v;
+            });
+            
+
+            const r4 = Try.sequence([r, r2, r3]);
+            const start = performance.now();
+            await expect(r4.get()).resolves.toEqual([4, 3, "2"]);
+            expect(r4.isSuccess()).toBe(true);
+            const timeTaken = (performance.now() - start) / 1000;
+            expect(timeTaken).toBeLessThan(5); //The maximum amount of time it should take is 5 seconds, because the longest running task is 4 seconds. The other 1 second is for the overhead of the test.
+        });
+
+        test("sequence should run all Try instances in parallel and results in Failure for the first instance that is a Failure", async () => {
+            const r = Try.success(4).map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000 * v));
+                return v;
+            });
+            const r2 = Try.success(3).map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000  * v));
+                return v;
+            });
+            const r3 = Try.success("2")
+            .map(async v => {
+                await new Promise(resolve => setTimeout(resolve, 1000  * Number(v)));
+                throw new Error("r3 error");
+                return v;
+            });
+            
+
+            const r4 = Try.sequence([r, r2, r3]);
+            const start = performance.now();
+            await expect(r4.get()).rejects.toThrow("r3 error");
+            expect(r4.isFailure()).toBe(true);
+            const timeTaken = (performance.now() - start) / 1000;
+            expect(timeTaken).toBeLessThan(3); //The maximum amount of time it should take is 3 seconds, because the longest running task until the error is thrown is 2 seconds. The other 1 second is for the overhead of the test.
+        });
+
+
     });
 });